FB - Add top bar to signed-in pages
Adds TopBar (nav links + logout) and UserLayout (auth guard + layout wrapper) components. Signed-in routes are now nested under UserLayout in the router, and the dashboard page is stripped of its now-redundant auth check and logout button.

Closes #69
