+++ frontend/src/main.tsx
@@ -8,17 +8,20 @@ import SignupPage from "./user/signup.page.tsx";
 import EditorPage from "./codeEditor/Editor.page.tsx";
 import FileBrowserPage from "./dashboard/FileBrowser.page.tsx";
 import Dashboard from "./user/dashboard.page.tsx";
+import UserLayout from "./components/UserLayout.tsx";
 
 createRoot(document.getElementById("root")!).render(
 	<StrictMode>
 		<BrowserRouter>
 			<Routes>
 				<Route path="/" element={<App />} />
-				<Route path="/edit/:fileId" element={<EditorPage />} />
 				<Route path="/signup" element={<SignupPage />} />
 				<Route path="/login" element={<LoginPage />} />
-				<Route path="/dashboard" element={<Dashboard />} />
-				<Route path="/filebrowser" element={<FileBrowserPage />} />
+				<Route element={<UserLayout />}>
+					<Route path="/dashboard" element={<Dashboard />} />
+					<Route path="/filebrowser" element={<FileBrowserPage />} />
+					<Route path="/edit/:fileId" element={<EditorPage />} />
+				</Route>
 			</Routes>
 		</BrowserRouter>
 	</StrictMode>,
